feat: ./test for integration tests, ./mock for user service mocks, unit test next to original file
