+++ backend/internal/controller/user_handler_test.go
@@ -1,13 +1,12 @@
-package test
+package controller
 
 import (
 	"bytes"
 	"encoding/json"
 	"net/http"
 	"net/http/httptest"
-	"sigma-test/internal/app"
-	"sigma-test/internal/controller"
 	"sigma-test/internal/entity"
+	"sigma-test/internal/mock"
 	"sigma-test/internal/util"
 	"testing"
 
@@ -20,8 +19,8 @@ func TestSignupWithMocks(t *testing.T) {
 	t.Run("should return 200 on successful signup", func(t *testing.T) {
 		r := gin.New()
 
-		userService := &MockUserService{MockDB: nil}
-		controller.InitUserHandler(r, userService)
+		userService := &mock.MockUserService{MockDB: nil}
+		InitUserHandler(r, userService)
 
 		requestBody := `{"email": "test@example.com", "password": "password", "role": "user"}`
 		body := bytes.NewBufferString(requestBody)
@@ -42,8 +41,8 @@ func TestSignupWithMocks(t *testing.T) {
 	t.Run("should return 400 on invalid body", func(t *testing.T) {
 		r := gin.New()
 
-		userService := &MockUserService{MockDB: nil}
-		controller.InitUserHandler(r, userService)
+		userService := &mock.MockUserService{MockDB: nil}
+		InitUserHandler(r, userService)
 
 		req, err := http.NewRequest("POST", "/api/signup", nil)
 		if err != nil {
@@ -63,8 +62,8 @@ func TestSignupWithMocks(t *testing.T) {
 
 		// Emutaling real users in database
 		mockUser := entity.User{ID: "test", Email: "test@test.com", Password: "test123", Role: "user"}
-		userService := &MockUserService{MockDB: []entity.User{mockUser}}
-		controller.InitUserHandler(r, userService)
+		userService := &mock.MockUserService{MockDB: []entity.User{mockUser}}
+		InitUserHandler(r, userService)
 
 		requestBody, _ := json.Marshal(mockUser)
 		body := bytes.NewBuffer(requestBody)
@@ -79,55 +78,7 @@ func TestSignupWithMocks(t *testing.T) {
 		r.ServeHTTP(res, req)
 
 		assert.Equal(t, http.StatusUnprocessableEntity, res.Code)
-	})
-}
 
-// Testing signup handler with SetupRouter
-func TestSignupIntegration(t *testing.T) {
-	t.Run("should return 200 on successful signup", func(t *testing.T) {
-		router := app.SetupRouter()
-		res := httptest.NewRecorder()
-
-		requestBody := []byte(`{"email": "test", "password": "test", "role": "user"}`)
-		body := bytes.NewReader(requestBody)
-		req, err := http.NewRequest("POST", "/api/signup", body)
-		if err != nil {
-			require.NoError(t, err)
-		}
-
-		router.ServeHTTP(res, req)
-
-		assert.Equal(t, http.StatusOK, res.Code)
-	})
-
-	t.Run("should return 400 on invalid body", func(t *testing.T) {
-		router := app.SetupRouter()
-		res := httptest.NewRecorder()
-
-		req, err := http.NewRequest("POST", "/api/signup", nil)
-		if err != nil {
-			require.NoError(t, err)
-		}
-
-		router.ServeHTTP(res, req)
-
-		assert.Equal(t, http.StatusBadRequest, res.Code)
-	})
-
-	t.Run("should return 422 if user exists", func(t *testing.T) {
-		router := app.SetupRouter()
-		res := httptest.NewRecorder()
-
-		jsonBody := []byte(`{"email": "test@test.com", "password": "test123", "role": "user"}`)
-		bodyReader := bytes.NewReader(jsonBody)
-		req, err := http.NewRequest("POST", "/api/signup", bodyReader)
-		if err != nil {
-			require.NoError(t, err)
-		}
-
-		router.ServeHTTP(res, req)
-
-		assert.Equal(t, http.StatusUnprocessableEntity, res.Code)
 	})
 }
 
@@ -136,8 +87,8 @@ func TestLoginWithMocks(t *testing.T) {
 		r := gin.New()
 
 		mockUser := entity.User{ID: "test", Email: "test@test.com", Password: "test", Role: "user"}
-		userService := &MockUserService{MockDB: []entity.User{mockUser}}
-		controller.InitUserHandler(r, userService)
+		userService := &mock.MockUserService{MockDB: []entity.User{mockUser}}
+		InitUserHandler(r, userService)
 
 		requestBody, _ := json.Marshal(mockUser)
 		body := bytes.NewBuffer(requestBody)
@@ -155,8 +106,8 @@ func TestLoginWithMocks(t *testing.T) {
 	t.Run("should return 400 on invalid body", func(t *testing.T) {
 		r := gin.New()
 
-		userService := &MockUserService{MockDB: nil}
-		controller.InitUserHandler(r, userService)
+		userService := &mock.MockUserService{MockDB: nil}
+		InitUserHandler(r, userService)
 
 		req, err := http.NewRequest("POST", "/api/login", nil)
 		if err != nil {
@@ -174,8 +125,8 @@ func TestLoginWithMocks(t *testing.T) {
 
 		// No users in mocked database, login impossible
 		mockUser := entity.User{}
-		userService := &MockUserService{MockDB: nil}
-		controller.InitUserHandler(r, userService)
+		userService := &mock.MockUserService{MockDB: nil}
+		InitUserHandler(r, userService)
 
 		requestBody, _ := json.Marshal(mockUser)
 		body := bytes.NewBuffer(requestBody)
@@ -199,8 +150,8 @@ func TestMeWithMocks(t *testing.T) {
 		mockUser := entity.User{ID: "test", Email: "test@test.com", Password: "test", Role: "user"}
 		mockToken, _ := util.GenerateJWTToken(mockUser.ID, mockUser.Role)
 
-		userService := &MockUserService{MockDB: []entity.User{mockUser}}
-		controller.InitUserHandler(r, userService)
+		userService := &mock.MockUserService{MockDB: []entity.User{mockUser}}
+		InitUserHandler(r, userService)
 
 		req, err := http.NewRequest("GET", "/api/me", nil)
 		if err != nil {
@@ -221,8 +172,8 @@ func TestMeWithMocks(t *testing.T) {
 		mockUser := entity.User{ID: "test", Email: "test@test.com", Password: "test", Role: "user"}
 		mockToken, _ := util.GenerateJWTToken("INVALID_ID", "INVALID_ROLE")
 
-		userService := &MockUserService{MockDB: []entity.User{mockUser}}
-		controller.InitUserHandler(r, userService)
+		userService := &mock.MockUserService{MockDB: []entity.User{mockUser}}
+		InitUserHandler(r, userService)
 
 		req, err := http.NewRequest("GET", "/api/me", nil)
 		if err != nil {
@@ -242,8 +193,8 @@ func TestMeWithMocks(t *testing.T) {
 
 		mockUser := entity.User{ID: "test", Email: "test@test.com", Password: "test", Role: "user"}
 
-		userService := &MockUserService{MockDB: []entity.User{mockUser}}
-		controller.InitUserHandler(r, userService)
+		userService := &mock.MockUserService{MockDB: []entity.User{mockUser}}
+		InitUserHandler(r, userService)
 
 		req, err := http.NewRequest("GET", "/api/me", nil)
 		if err != nil {
@@ -255,21 +206,5 @@ func TestMeWithMocks(t *testing.T) {
 
 		t.Log(res.Body.String())
 		assert.Equal(t, http.StatusUnauthorized, res.Code)
-	})
-}
-
-func TestPingPongIntegration(t *testing.T) {
-	t.Run("should return ping on pong", func(t *testing.T) {
-		router := app.SetupRouter()
-		res := httptest.NewRecorder()
-		req, err := http.NewRequest("GET", "/ping", nil)
-		if err != nil {
-			require.NoError(t, err)
-		}
-
-		router.ServeHTTP(res, req)
-
-		assert.Equal(t, http.StatusOK, res.Code)
-		assert.Equal(t, `{"message":"pong"}`, res.Body.String())
 	})
 }
