+++ backend/internal/mock/user_service_mock.go
@@ -1,4 +1,4 @@
-package test
+package mock
 
 import (
 	"errors"
